Corrigido BUG ao listar Template Avaliação.
