+++ src/main/java/com/infnet/avaliacao/dto/impl/TemplateTopicoDTO.java
@@ -16,8 +16,6 @@ public class TemplateTopicoDTO implements IDTO<TemplateTopico> {
 
     private String enunciado;
 
-    private List<TemplateAvaliacaoTopicoPerguntaDTO> templateAvaliacaoTopicoPerguntaDTOList;
-
     /**
      * {@inheritDoc}
      */
@@ -26,9 +24,6 @@ public class TemplateTopicoDTO implements IDTO<TemplateTopico> {
         TemplateTopico templateTopico = new TemplateTopico();
         templateTopico.setId(this.getId());
         templateTopico.setEnunciado(this.getEnunciado());
-        templateTopico.setTemplateAvaliacaoTopicoPerguntaList(
-                TemplateAvaliacaoTopicoPerguntaDTO.convertListDtoToListEntity(
-                        this.getTemplateAvaliacaoTopicoPerguntaDTOList()));
         return templateTopico;
     }
 
@@ -42,9 +37,6 @@ public class TemplateTopicoDTO implements IDTO<TemplateTopico> {
         TemplateTopicoDTO templateTopicoDTO = new TemplateTopicoDTO();
         templateTopicoDTO.setId(templateTopico.getId());
         templateTopicoDTO.setEnunciado(templateTopico.getEnunciado());
-        templateTopicoDTO.setTemplateAvaliacaoTopicoPerguntaDTOList(
-                TemplateAvaliacaoTopicoPerguntaDTO.convertListEntityToListDto(
-                        templateTopico.getTemplateAvaliacaoTopicoPerguntaList()));
         return templateTopicoDTO;
     }
 
@@ -98,11 +90,4 @@ public class TemplateTopicoDTO implements IDTO<TemplateTopico> {
         this.enunciado = enunciado;
     }
 
-    public List<TemplateAvaliacaoTopicoPerguntaDTO> getTemplateAvaliacaoTopicoPerguntaDTOList() {
-        return templateAvaliacaoTopicoPerguntaDTOList;
-    }
-
-    public void setTemplateAvaliacaoTopicoPerguntaDTOList(List<TemplateAvaliacaoTopicoPerguntaDTO> templateAvaliacaoTopicoPerguntaDTOList) {
-        this.templateAvaliacaoTopicoPerguntaDTOList = templateAvaliacaoTopicoPerguntaDTOList;
-    }
 }
